found one of the live data issues and added a warning
also made a little progress towards organisation

a basic check on data from the same distributions seems to fail to solve for multiplying by 1.0, probably a clear debuggable error, haven't written a test for it yet.

it's also notable that multiplying by a scalar < 1 seems to sometimes find an answer
that is a little greater than the correct one, maybe due to randomness at the edges
of a signal shrinking how large its data makes it look, dunno.