organised a smidge more
the biggest present issue appears to be solving for results that are precise, and are
not on the midpoint of a bucket.  a zero vector is output since the buckets aren't integrated.

the most direct approach to move towards solving this appears to be to consider the regions
covered between the bucket edges.  this leaves out maxima and minima points, but covers most situations.  additionally densities will be wrong if functions are not roughly linear between buckets.

one approach for maxima/minima is to use an iterative algorithm.  this opens up the idea
of using an iterative algorithm to also find the most probable solution.  that's an orthogonal
concern, but shares concept.
    iterative algorithms are in line with a simple technique called adaptive subsampling that
    could be used to output a visual chart similar to a histogram (buckets would be variably-
    sized).
        the architecture is presently pre-generated, so it could be helpful to reuse previous
        architecture data if doing this.
another approach is to ask the user to provide functions that give the minima and maxima
another approach is to ask the user to provide functions that give proper integrals
of the expressions for the actual bucket densities.  this would be the most accurate.
    such integrals could eventually be automatically produced via analytical solving in python,
    by passing symbolic numbers as parameters to the expression.
    likely a library to do this already exists.